requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.1
aiosqlite==0.20.0
torch==2.2.2
transformers==4.41.2
bitsandbytes==0.43.1
accelerate==0.31.0
peft==0.11.1
trl==0.9.4
datasets==2.20.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
slowapi==0.1.9
httpx==0.27.0
motor==3.4.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
pydantic[email]==2.7.4
pytest==8.2.2
pytest-asyncio==0.23.7
